Remove tests for invalid method
Turns out reqwest also accepts any request method and
dutifully makes a request to the server.
The servers might respond with an HTTP status code > 2xx,
but apart from that, these requests are totally valid.

I don't know why this was done, but I assume because
of compatibility with potential future versions of the
HTTP spec. In any case, I'm planning to lift the
restrictrions on accepting only `GET` and `HEAD` requests
because of this and removing the test.