Cleanup Docker image
Move `lychee` binary to `/usr/local/bin` to make it callable
without a full path.
This makes it easier to use from the lychee Github action
in the future.
Also the user setup is making things unnecessary complicated
and could be prohibited in some environments so this commit
is removing it and uses the defaults.