Cache `create_uri_from_path` to decrease allocs
While profiling local file handling, I noticed that resolving paths was taking a
significant amount of time. It also caused quite a few allocations.
By caching the path and using a constant value for the current
directory, we can reduce the number of allocs by quite a lot.
For example, when testing on the sentry documentation, we now do 50,4%
less allocations now in total. That's just a single test-case of course,
but it's probably also helping for many other cases.