Skip Docker login on PRs from forks (#586)
Since github.repository_owner refers to the base repository on pull request events, while secrets need to exist on the head repository, the Docker Hub login fails for pull requests opened from forks. This commit assures that this step in case of pull request events only runs for internal pull requests, i.e. when head and base repository are the same.

For the actual Docker Hub upload, no change is required: The build can run in every case, the upload is not done on pull request events in general.

Signed-off-by: MichaIng <micha@dietpi.com>