Add support for dumping inputs
Dumps all detected inputs to stdout without checking them
This is useful for debugging purposes.

The implementation is suboptimal at the moment.
- The collected inputs get read from disk before being written to stdout.
- If inputs don't contain any links, they are ignored.

It would be better to just stream the inputs and not load them into memory.
This would require some refactoring in the `lychee-lib` crate,
so it is left as a TODO for now.