Fix nested URL extraction in verbatim elements
Skipping URLs in verbatim elements didn't take nested
elements into consideration, which are not verbatim.

For instance, the following HTML snippet would yield
`https://example.com` in non-verbatim mode, even if
it is nested inside a verbatim `<pre>` element:

```
<pre><a href="https://example.com">link</a></pre>
```

This commit fixes the behavior for both `html5gum` and
`html5ever`.

Note that nested verbatim of the same type still are not
handled correctly.

For instance,  the following HTML snippet would still yield
`https://example.com`:

```
<pre><pre></pre><a href="https://example.com">link</a></pre>
```

The reason is that we currently only keep track of a single
verbatim element and not a stack of elements, which we
would need to unwind to resolve the situation.

Fixes https://github.com/lycheeverse/lychee/issues/986