Fix tests
Okay this was quite a dig. Clap uses the default value not as is, but will
use `to_string` (from the Display impl) to convert the default value into
a string and then re-parses it via the `FromStr` trait. The `Display` impl
from `AcceptSelector` included surrounding square brackets `[ ]` which failed
to parse.

cargo expand really helped here, as I had to dig into the code generated by the Clap `Parser` derive macro.

I'm still unsure why Clap handles default values in this way, but there surely is a reason for it. I'm just too lazy to do further research into that.