fix: Bring back error output for links
With the last lychee release, we simplified the status output for links.  While
this reduced the visual noise, it also accidentially caused the source of
errors to not be printed anymore. This change brings back the additional error
information as part of the final report output. Furthermore, it shows the error
information in the progress output if verbose mode is activated.