tests: refactor `test_fragments` to clarify expected successes/fails
this PR improves `test_fragments` by separating the URLs into separate
lists for expected failures and expected successes. each of these are
searched in stdout and/or stderr to check that the URL has the expected
outcome. this makes the expected test behaviour much more obvious, where
it was previously only implied by the "x total" and "x OK" strings.

this PR makes no changes to actual behaviour. hopefully, it will make it
easier to test future changes to the fragment behaviour.

previously, the test only checked that stderr contains a list of URLs.
because `--verbose` is specified, stderr would contain both successful and
failing URLs, so this didn't really test much. it only tested that the
URLs were detected by lychee.

also, search terms are now suffixed with whitespace to prevent `file.html`
from matching `file.html#something`. the test should be consider these as
different links, but the previous code had the potential to match
`file.html#something` when looking for `file.html`.

in making this PR, we find that a fragment link to an empty file is
always treated as a valid link because it is detected as a plaintext
file. maybe there should be a warning when a URL points to a plaintext
file and a fragment check is requested
([relevant code](https://github.com/lycheeverse/lychee/blob/ea415c8db4597c383f88b5fc9894fd86a7245494/lychee-lib/src/utils/fragment_checker.rs#L133)).