add warning when checking fragments in a plaintext file
for files detected as Plaintext type, fragment checking is not supported
and will always return true. at the moment, this is done silently and
this might be surprising if the user has requested fragment checking.
note that the Plaintext case includes empty files and unknown file
types, since Plaintext is used as the fallback file type.

this change adds a warning to fragment_checker when this case is
reached.

the other reasonable place to put this warning would be in
`lychee-lib/src/checker/file.rs` next to the other "Skipped fragment
check" warnings. i have decided to put it in `fragment_checker.rs` so
that it is next to the `return Ok(true)` statement. this will make it
easier in case the Plaintext case is ever changed.