tests: refactor `test_fragments` to clarify expected successes/fails (#1776)
* tests: refactor `test_fragments` to clarify expected successes/fails

this PR improves `test_fragments` by separating the URLs into separate
lists for expected failures and expected successes. each of these are
searched in stdout and/or stderr to check that the URL has the expected
outcome. this makes the expected test behaviour much more obvious, where
it was previously only implied by the "x total" and "x OK" strings.

this PR makes no changes to actual behaviour. hopefully, it will make it
easier to test future changes to the fragment behaviour.

previously, the test only checked that stderr contains a list of URLs.
because `--verbose` is specified, stderr would contain both successful and
failing URLs, so this didn't really test much. it only tested that the
URLs were detected by lychee.

also, search terms are now suffixed with whitespace to prevent `file.html`
from matching `file.html#something`. the test should be consider these as
different links, but the previous code had the potential to match
`file.html#something` when looking for `file.html`.

in making this PR, we find that a fragment link to an empty file is
always treated as a valid link because it is detected as a plaintext
file. maybe there should be a warning when a URL points to a plaintext
file and a fragment check is requested
([relevant code](https://github.com/lycheeverse/lychee/blob/ea415c8db4597c383f88b5fc9894fd86a7245494/lychee-lib/src/utils/fragment_checker.rs#L133)).

* borrow in for loop

* add remaining URLs to classify all detected URLs

also update outdated text in file1.md

* use .len() to compute expected counts

* move total check last so the more informative ok/err checks are first

* fix text which wasn't updated after revert

* add warning when checking fragments in a plaintext file

for files detected as Plaintext type, fragment checking is not supported
and will always return true. at the moment, this is done silently and
this might be surprising if the user has requested fragment checking.
note that the Plaintext case includes empty files and unknown file
types, since Plaintext is used as the fallback file type.

this change adds a warning to fragment_checker when this case is
reached.

the other reasonable place to put this warning would be in
`lychee-lib/src/checker/file.rs` next to the other "Skipped fragment
check" warnings. i have decided to put it in `fragment_checker.rs` so
that it is next to the `return Ok(true)` statement. this will make it
easier in case the Plaintext case is ever changed.

* use info level for plaintext fragment messge