review: make `index_files` an Option for clearer default behavior
this changes all the `index_files` fields to be `Option<Vec<String>>`
instead of only a vec. this lets API users simply pass a None to
get the default behaviour.

documentation at the higher level interfaces (e.g., the CLI help)
is changed to remove reference to `.` as the default behaviour.
now, the default behaviour is defined in terms of its observable
behaviour without reference to this implementation detail.

that said, the index file resolution still has this special case
that `.` is allowed in the `index_files` list. this allows the
(possibly esoteric) case where a user passes `--index-files index.html,.`
to use an index file if it exists but still accept the link regardless.

the None case for index_files is still implemented by mapping
to an index file list of `["."]`, but this is now only an
implementation detail of `apply_index_files`.

finally, using an Option is nicer than special-casing the empty
`Vec<String>` case, because the empty vec case /should/ reject
all directory links. this follows logically because the fewer
items in the index file list, the fewer directory links are valid.