fix tests bug, where an assertion was weaker than desired
this assertion is meant to check that there are two failing empty dir links. however, by checking stderr, it would've passed even if those links succeeded. this is because `--verbose` prints all links (both valid and invalid) to stderr.

we change it to search in stdout which contains only the failing links.