clarify `--base-url` and `--root-dir` and their interactions
this more clearly documents the behaviour of these flags. this makes no
changes to behavior. as always, this documents the usage as i observe
right now and does not presume that this behaviour is good or correct.

i also hope to clarify some more subtle points:
- the base-url is uniformly applied to every local file being checked,
  regardless of the filesystem path of each local file. this means that
  a relative link `a/b` resolves to exactly the same URL no matter which
  local file it appears in.
- base-url is affected by trailing slashes.
- define behavior when root-dir and base-url are both specified. namely,
  root-dir becomes a subpath of the base-url's domain.

honestly, the more i think about it, the more i think this behavior is
wrong or has diverged from its original intention. for instance, let's
look at [the docs recipe about `--root-dir`][1]:

[1]: https://lychee.cli.rs/recipes/root-dir/#using-both-together

> Sometimes you need both:
>
> ```
> lychee \
>   --root-dir "$(pwd)/public" \
>   --base-url https://example.com/ \
>   "public/**/*.html"
> ```
>
> This tells lychee:
>
> 1. Look for /-prefixed files in ./public/
> 2. Resolve relative links against https://example.com/

with the above points in mind, this example is nonsensical. firstly, the
root-dir will make absolute links become something like
`https://example.com/path/to/pwd/public/absolute-link` which is unlikely
to exist on the remote host. this is in contradiction with point (1) in
the quote. secondly, /every/ relative link is resolved against exactly
`https://example.com/`, without considering its position relative to
root-dir. this seems wrong too.

but in any case, this PR clarifies the behavior as it is right now.