add `--index-files` flag, and turn off index file checking by default (#1777)
* add index_files option

* fix compilation

* phrasing

* fjdisaofjdisoajfidosa

i'm thinking about this more than anyone ever has. index_files interacts
with fallback_extensions in unclear ways. i feel like we should just
collapse them both and have like

suffixes = ["", ".html", "/index.html"]

in conjunction with a "--accept-directories" flag that defaults to true.

this suffixes list would be used to implement the fallback_extensions.

* early switch between fallback extensions and index files

* docs and it compiles now

* COW

* update readme text

* increment errors due to not searching index.html by default

now fails:

fixtures/fragments/sub_dir#a-link-inside-index-html-inside-sub-dir

* tests and fixtures

* fmt and fix test

the test was broken because we no longer check index.html by default.
i change the test URLs to have `/index`. this still tests that the
fallback extension `.html` is being applied correctly.

* lint

* add more corner cases to tests

* separate test cases for corner cases

* fix index_files default in ClientBuilder (oops!)

* tests: now expected to fail `#a-link-inside-index-html-inside-sub-dir`

because we no longer check index.html by default

* revert README.md change where we hardcoded the version

* review: make `index_files` an Option for clearer default behavior

this changes all the `index_files` fields to be `Option<Vec<String>>`
instead of only a vec. this lets API users simply pass a None to
get the default behaviour.

documentation at the higher level interfaces (e.g., the CLI help)
is changed to remove reference to `.` as the default behaviour.
now, the default behaviour is defined in terms of its observable
behaviour without reference to this implementation detail.

that said, the index file resolution still has this special case
that `.` is allowed in the `index_files` list. this allows the
(possibly esoteric) case where a user passes `--index-files index.html,.`
to use an index file if it exists but still accept the link regardless.

the None case for index_files is still implemented by mapping
to an index file list of `["."]`, but this is now only an
implementation detail of `apply_index_files`.

finally, using an Option is nicer than special-casing the empty
`Vec<String>` case, because the empty vec case /should/ reject
all directory links. this follows logically because the fewer
items in the index file list, the fewer directory links are valid.

* docs: add `--index-files ''` and `--index-files index.html,.`

as suggested.

* tests: add cli tests for `--index-files`

this makes sure that the 3 examples in the --help message
function as expected. in particular, this covers the case of
`--index-files ''` which might easily be inadvertently broken.

* review: confirm path traversal is allowed in `--index-files`

as suggested.

* review: typo

Co-authored-by: Thomas Zahner <thomas.zahner@protonmail.ch>

* fix tests bug, where an assertion was weaker than desired

this assertion is meant to check that there are two failing empty dir links. however, by checking stderr, it would've passed even if those links succeeded. this is because `--verbose` prints all links (both valid and invalid) to stderr.

we change it to search in stdout which contains only the failing links.

---------

Co-authored-by: Thomas Zahner <thomas.zahner@protonmail.ch>