apply fallback extensions if directory links resolve to a dir (#1826)
* add test case demonstrating 1817

https://www.github.com/lycheeverse/lychee/issues/1817

* apply fallback extensions if directory links resolve to a directory

this restores the 0.19 behaviour when a directory `same_name` is
next to a file `same_name.html`, a local link to `same_name` with
fallback extension `html` will again be able to resolve to
`same_name.html`.

probably, this always makes sense because directories are treated
as having no fragments, so resolving to real file with a fallback
extension (if it exists) will potentially contain more fragments
and thus be "more useful".

NOTE: this is probably risky. because, fallback extensions should only
be applied to relative links which are /not/ ending in `/`. links
ending in `/` should be required to resolve to directories (or their
index files). unfortunately, i think the detail of the trailing `/`
is normalised away fairly early in the pipeline. idk what to do here,
maybe we just fix this issue for now and file a bug about the trailing
slash.

* fmt