chore(file checker): separate resolving and checking for easier testing
this does a simple refactor to break up the `check_path` function.
previously, this had the responsibility of resolving index
files/fallback extensions _and_ checking fragments. this was done in one
step which made it hard to test - in order to test the resolving, we had
to test it indirectly by checking fragments in certain files.

this introduces a new `resolve_local_path` function which only applies
the index file + extensions, so we can use this in unit tests and make
assertions about the resolved path.

this should have no changes to any behaviour.