Redirect reporting (#1725)
Fixes #1433

We want to report redirects at the end of the link check process, the same way we do for example with `--suggest`. The output format `markdown` for example should then also contain a section with all listed redirects.

**Summary bug**

There is a category `Redirected` in the summary. However, redirects were never counted. This is fixed with 6b28e2b.
This means `echo "https://http.codes/301 " | cargo run - --format detailed` yielded

```
📝 Summary
---------------------
🔍 Total............1
✅ Successful.......1
⏳ Timeouts.........0
🔀 Redirected.......0
👻 Excluded.........0
❓ Unknown..........0
🚫 Errors...........0
```

now it yields

```
📝 Summary
---------------------
🔍 Total............1
✅ Successful.......0
⏳ Timeouts.........0
🔀 Redirected.......1
👻 Excluded.........0
❓ Unknown..........0
🚫 Errors...........0
```

---------

Co-authored-by: Matthias Endler <matthias@endler.dev>