move CLI-related suggestions into .context in lychee-bin
this does have the side-effect of attaching the --help and --root-dir
suggestions even to the "cannot be a base" error.

```
error: invalid value 'a:datafdajsio' for '--base-url <BASE_URL>': Error
with base dir `a:datafdajsio` : The given URL cannot be used as a base
URL. See `--help` for more information. If you want to resolve
root-relative links in local files, also see `--root-dir`.
```

this could be a bit confusing, but idk a way around it without
inspecting the error message of InvalidBase to determine which context
to add.