feat: improve InvalidIndexFile error to mention index file names (#1859)
* feat: improve InvalidIndexFile error to mention index file names

this makes it clearer which files are causing the link check to fail.
it also lets us customise the message in the special case of
`--index-files ''`, which could otherwise be confusing with a generic
message.

the messages are:
```
[ERROR] file:///h | Cannot find index file within directory: An index file (index.html) is required
[ERROR] file:///h | Cannot find index file within directory: An index file (index.html, or index.htm) is required
[ERROR] file:///h | Cannot find index file within directory: An index file (index.html, index.htm, or index.md) is required
```

and for the `--index-files ''` case:
```
[ERROR] file:///h | Cannot find index file within directory: No directory links are allowed because index_files is defined and empty
```

i do wonder if this last error message is still confusing because of the
"Cannot find index file within directory" text. maybe it should be
changed to a more general "Cannot resolve link to local directory"?

for comparison, the old message was always:
```
[ERROR] file:///h | Cannot find index file within directory: Index file not found in directory. Check if index.html or other index files exist
```

* move dropping into file.rs

* add some tests

* revert spurious