feat: print warning if input source matches no files
it looks like this:
```
$ lychee 'non-existing/*' '*.fdsamifdsa' 'no-matches?????' empty-dir
   [WARN ] *.fdsamifdsa: No files found for this input source
   [WARN ] no-matches?????: No files found for this input source
   [WARN ] non-existing/*: No files found for this input source
   [WARN ] empty-dir: No files found for this input source
  0/0 ━━━━━━━━━━━━━━━━━━━━ Finished extracting links                                                                                                                                        🔍 0 Total (in 0s) ✅ 0 OK 🚫 0 Errors
```

this is implemented as a sneaky `log::warn!` which is probably
undesirable. however, the code isn't set up yet to properly pass errors
from the input resolving stage up to the top level.

even if it was, the "error on empty input" error should probably be put
behind a flag. it would also be tricky because the reporting expects a
ResolvedInputSource rather than InputSource, and these warnings arise in
the process of resolving. (maybe we could slightly wind back #1840 to
make this work)

anyway, much to think about.