fix: inverted gitignore behaviour for --dump-inputs
the use of `no_ignore` was missing a `!` before being passed as
skip_ignored. i figure there was some confusion about the meaning of
"no ignore", it can be either:

- skip ignored files, or
- skip the ignore process itself.

the actual meaning is the second one. the no_ignore value was used
inconsistently between the main logic and the dump-inputs logic.

related to https://www.github.com/lycheeverse/lychee/issues/1767